+++ src/main/kotlin/com/alefranc/composescreenshotplugin/content/PluginTexts.kt
@@ -1,17 +1,17 @@
 package com.alefranc.composescreenshotplugin.content
 
 object PluginTexts {
-    const val ACTION_TEXT_DELETE_REFERENCES = "Delete references images"
-    const val ACTION_TEXT_OPEN_REPORT = "Open report"
+    const val ACTION_TEXT_DELETE_REFERENCES = "Delete Reference Images"
+    const val ACTION_TEXT_OPEN_REPORT = "Open Report"
     const val ACTION_TEXT_RECORD = "Record"
-    const val ACTION_TEXT_SHOW_REFERENCES = "Show references images"
+    const val ACTION_TEXT_SHOW_REFERENCES = "Show Reference Images"
     const val ACTION_TEXT_VERIFY = "Verify"
 
-    const val GUTTER_RUN_SCREENSHOT_TEST = "Screenshot test"
-    const val GUTTER_RUN_SCREENSHOT_TESTS = "Screenshot test(s)"
+    const val GUTTER_RUN_SCREENSHOT_TEST = "Screenshot Test"
+    const val GUTTER_RUN_SCREENSHOT_TESTS = "Screenshot Test(s)"
 
     const val NOTIFICATION_VERIFY_TEXT = "Screenshot testing finished"
-    const val NOTIFICATION_IMAGES_DELETED_TEXT = "References images deleted"
+    const val NOTIFICATION_IMAGES_DELETED_TEXT = "Reference images deleted"
 
     const val ERROR_TITLE = "Error"
     const val ERROR_NO_FILES_FOUND = "No files found to open"